metrics api test
