+++ rundeckapp/src/test/groovy/rundeck/controllers/ExecutionControllerTests.groovy
@@ -21,6 +21,7 @@ import com.dtolabs.rundeck.app.internal.logging.RundeckLogFormat
 import com.dtolabs.rundeck.app.support.ExecutionQuery
 import grails.test.mixin.Mock
 import grails.test.mixin.TestFor
+import grails.testing.gorm.DataTest
 import groovy.json.JsonSlurper
 import groovy.mock.interceptor.MockFor
 import org.quartz.JobExecutionContext
@@ -31,6 +32,8 @@ import rundeck.Workflow
 import rundeck.services.*
 import rundeck.services.logging.ExecutionLogState
 import rundeck.services.logging.WorkflowStateFileLoader
+
+import java.sql.Time
 
 import static org.junit.Assert.assertEquals
 import static org.junit.Assert.assertNotNull
@@ -655,44 +658,42 @@ class ExecutionControllerTests  {
      * Test metrics calculations.
      */
     public void testApiExecutionsMetrics() {
+
         def controller = new ExecutionController()
 
-        def execs = createTestExecs()
-
-//        def fwkControl = new MockFor(FrameworkService, false)
-//        fwkControl.demand.getAuthContextForSubjectAndProject { subj, proj -> return null }
-//        fwkControl.demand.filterAuthorizedProjectExecutionsAll { framework, List<Execution> results, Collection actions ->
-//            return results
-//        }
-//        controller.frameworkService = fwkControl.proxyInstance()
         controller.request.api_version = 29
         controller.request.contentType = "application/json"
         controller.params.project = "Test"
-
-//        def execControl = new MockFor(ExecutionService, false)
-//        execControl.demand.queryExecutions { ExecutionQuery query, int offset, int max ->
-//            assert null != query
-//            assert "Test" == query.projFilter
-//            return [result: execs, total: execs.size()]
-//        }
-//        execControl.demand.respondExecutionsXml { request, response, List<Execution> execsx, paging ->
-//            return true
-//        }
-//        controller.executionService = execControl.proxyInstance()
 
         def apiMock = new MockFor(ApiService, false)
         apiMock.demand.requireVersion { request, response, int min ->
             assertEquals(29, min)
             return true
         }
-//        apiMock.demand.renderSuccessXml { request, response ->
-//            return true
-//        }
         controller.apiService = apiMock.proxyInstance()
+
+        // mock exec service
+        controller.executionService = new ExecutionService()
+
+        // Mock metrics criteria
+        mockDomain Execution
+        def metricCriteria = new Expando()
+        metricCriteria.get = { Closure c -> [
+                count      : 3,
+                durationMax: new Time(0, 9, 0),
+                durationMin: new Time(0, 2, 0),
+                durationSum: new Time(0, 15, 0)
+            ]
+        }
+        Execution.metaClass.static.createCriteria = { metricCriteria }
+
+        // Call controller
         controller.apiExecutionMetrics(new ExecutionQuery())
 
+        // Parse response.
         def resp = new JsonSlurper().parseText(response.text)
 
+        // Check respose.
         assert 200 == controller.response.status
         assert resp.total == 3
 //        assert resp.status.succeeded == 3
